pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "formless-agent-service"
version = "0.1.0"
description = "Dyno Formless Agent Service"
authors = [
  { name = "Luke Carpenter" }
]
requires-python = ">=3.10"

# ---------------------------------------------------------------------------
# Runtime dependencies (PRODUCTION)
# ---------------------------------------------------------------------------

dependencies = [
    # Core framework
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.25",
    "psycopg2-binary>=2.9.9",
    "pyliquibase>=1.4.1",
    "python-dotenv>=1.0.1",

    # JWT support
    "python-jose[cryptography]>=3.3.0",

    # -----------------------------------------------------------------------
    # NEW: AI & JSON Tools
    # -----------------------------------------------------------------------
    "cohere>=5.0.0",
    "jsonschema>=4.21.0",
    "json-repair>=0.15.0",
    "pystache>=0.6.0",

    # -----------------------------------------------------------------------
    # Messaging (Celery + RabbitMQ)
    # -----------------------------------------------------------------------
    "celery>=5.4.0",

    "json-logic>=0.6.3",

    # -----------------------------------------------------------------------
    # Pattern matching and fuzzy logic
    #
    # phonenumbers: parse and format international phone numbers to E.164.  If
    # unavailable, identity extraction gracefully skips phone parsing.
    # usaddress: parse US addresses into structured components (optional).
    # rapidfuzz: compute string similarity metrics (e.g. Levenshtein,
    # Jaro–Winkler) for fuzzy name matching during deduplication.
    # jellyfish: provides phonetic algorithms including Double Metaphone.
    # -----------------------------------------------------------------------
    "phonenumbers>=8.13.6",
    "usaddress>=0.5.11",
    "rapidfuzz>=3.2.0",
    "jellyfish>=1.0.0",

    "jinja2>=3.1.0,<4.0.0",
    "httpx>=0.27.0,<0.28.0",

    # -------------------------------------------------------------------
    # OpenTelemetry tracing (core API & SDK + instrumentation)
    # -------------------------------------------------------------------
    # Core OTel API and SDK: stable version 1.x provides the API and
    # implementation necessary to create spans and export them via OTLP.
    "opentelemetry-api>=1.24.0,<2.0.0",
    "opentelemetry-sdk>=1.24.0,<2.0.0",
    "opentelemetry-exporter-otlp>=1.24.0,<2.0.0",
    # Instrumentations: pre-1.0 releases use 0.x versions.  We pin
    # greater than or equal to 0.45b0 (first beta with httpx/Celery
    # support) and less than 1.0.0 to avoid breaking changes.
    "opentelemetry-instrumentation-fastapi>=0.45b0,<1.0.0",
    "opentelemetry-instrumentation-asgi>=0.45b0,<1.0.0",
    "opentelemetry-instrumentation-httpx>=0.45b0,<1.0.0",
    "opentelemetry-instrumentation-sqlalchemy>=0.45b0,<1.0.0",
    "opentelemetry-instrumentation-logging>=0.45b0,<1.0.0",
    "opentelemetry-instrumentation-celery>=0.45b0,<1.0.0",
]


# ---------------------------------------------------------------------------
# Optional dependency groups
# ---------------------------------------------------------------------------
[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.setuptools]
package-dir = {"" = "app"}

[tool.setuptools.packages.find]
where = ["app"]

# ---------------------------------------------------------------------------
# Tooling configs
# ---------------------------------------------------------------------------

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
